ml/requirements.txt
# Python Dependencies for Haemologix ML

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Data processing
pyyaml>=6.0

# Optional: TensorBoard for logging
tensorboard>=2.14.0

# Database (if using direct DB access)
# psycopg2-binary>=2.9.0  # Uncomment if needed

